Share as much Bitmap code with Windows as possible
